+++ pkg/providers/gtm/testdata/TestResGtmProperty/create_multiple_traffic_targets.tf
@@ -18,7 +18,6 @@ resource "akamai_gtm_property" "tfexample_prop_1" {
     enabled       = true
     weight        = 200
     servers       = ["1.2.3.4"]
-    name          = "test1"
     handout_cname = "test"
   }
 
@@ -27,7 +26,6 @@ resource "akamai_gtm_property" "tfexample_prop_1" {
     enabled       = true
     weight        = 200
     servers       = ["1.2.3.5"]
-    name          = "test2"
     handout_cname = "test"
   }
 
@@ -36,7 +34,6 @@ resource "akamai_gtm_property" "tfexample_prop_1" {
     enabled       = true
     weight        = 200
     servers       = ["1.2.3.6"]
-    name          = "test3"
     handout_cname = "test"
   }