+++ pkg/providers/gtm/testdata/TestResGtmProperty/servers/change_server_and_diff_traffic_target_order.tf
@@ -15,7 +15,6 @@ resource "akamai_gtm_property" "tfexample_prop_1" {
     enabled       = true
     weight        = 200
     servers       = ["1.2.3.6"]
-    name          = "test2"
     handout_cname = "test"
   }
 
@@ -24,7 +23,6 @@ resource "akamai_gtm_property" "tfexample_prop_1" {
     enabled       = true
     weight        = 200
     servers       = ["1.2.3.7", "1.2.3.8"]
-    name          = "test3"
     handout_cname = "test"
   }
 
@@ -33,7 +31,6 @@ resource "akamai_gtm_property" "tfexample_prop_1" {
     enabled       = true
     weight        = 200
     servers       = ["10.20.30.10", "1.2.3.5"]
-    name          = "test1"
     handout_cname = "test"
   }
   liveness_test {